backend/requirements.txt
langchain
langchain-cerebras
langchain-core
duckduckgo-search
wikipedia
llama-index-llms-cerebras
llama-index
cerebras_cloud_sdk
llama-index-callbacks-arize-phoenix
